crossplane-functions/consumers/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
idna==3.4
pathspec==0.11.1
platformdirs==2.6.2
PyYAML==6.0
tomli==2.0.1
argparse==1.4.0